refactor(client): refactoring helper function; add nav test
